+++ Source/ray_interpolated_projection.cu
@@ -109,9 +109,9 @@ __global__ void kernelPixelDetector( Geometry geo,
         Point3D uvOrigin,
         float maxdist){
     
-    int y = blockIdx.y * blockDim.y + threadIdx.y;
-    int x = blockIdx.x * blockDim.x + threadIdx.x;
-    size_t idx =  x  * geo.nDetecV + y;
+    unsigned long  y = blockIdx.y * blockDim.y + threadIdx.y;
+    unsigned long x = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long  idx =  x  * geo.nDetecV + y;
 
     if ((x>= geo.nDetecU) | (y>= geo.nDetecV))
         return;
@@ -226,7 +226,7 @@ int interpolation_projection(float const * const img, Geometry geo, float** resu
     
     Point3D source, deltaU, deltaV, uvOrigin;
     float maxdist;
-    for (int i=0;i<nalpha;i++){
+    for (unsigned int i=0;i<nalpha;i++){
         
         geo.alpha=alphas[i];
         //precomute distances for faster execution
